Catalogue, Inventory, and Order Endpoints

+++ src/main/java/com/usepipeline/portal/web/license/model/ActiveLicenseModel.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -18,7 +19,7 @@ public class ActiveLicenseModel extends LicenseModel {
         return new ActiveLicenseModel(licenseEntity.getLicenseId(), licenseEntity.getLicenseHash(), licenseEntity.getExpirationDate(), licenseEntity.getType(), licenseEntity.getAvailableLicenseSeats(), licenseEntity.getMaxLicenseSeats(), licenseEntity.getMonthlyCost(), organizationAccount);
     }
 
-    public ActiveLicenseModel(UUID licenseId, UUID licenseHash, LocalDate expirationDate, String type, Long availableLicenseSeats, Long maxLicenseSeats, Double monthlyCost, LicensedOrganizationAccountModel organizationAccount) {
+    public ActiveLicenseModel(UUID licenseId, UUID licenseHash, LocalDate expirationDate, String type, Long availableLicenseSeats, Long maxLicenseSeats, BigDecimal monthlyCost, LicensedOrganizationAccountModel organizationAccount) {
         super(licenseId, licenseHash, expirationDate, type, availableLicenseSeats, maxLicenseSeats, monthlyCost, true);
         this.organizationAccount = organizationAccount;
     }
