+++ src/main/java/com/usepipeline/portal/web/license/model/LicenseCreationRequestModel.java
@@ -5,13 +5,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class LicenseCreationRequestModel {
     private String type;
     private Long licenseSeats;
-    private Double monthlyCost;
+    private BigDecimal monthlyCost;
     private PortalDateModel expirationDate;
 
 }
